+++ common/templatetags/str_handle.py
@@ -17,3 +17,7 @@ def split(value: str, sep):
 @register.filter
 def index(value: list, idx):
     return value[idx]
+
+@register.simple_tag
+def retrieve_roster_name(path: str):
+    return path.split('/')[1]