+++ static/js/deploy.js
@@ -16,7 +16,6 @@ function salt(ele, url, successStr) {
     })
 }
 
-
 function importMinions(event) {
     successStr = 'Import Minions Success.'
     salt(this, event.data.url, successStr)
@@ -25,31 +24,6 @@ function importMinions(event) {
 function refreshMinionsHealth(event) {
     successStr = 'Refresh Minions Health Success.'
     salt(this, event.data.url, successStr)
-}
-
-// ajax: delete minion
-function deleteMinion() {
-    var tds = $(this).parent('td').siblings('td')
-    var hostname = tds[0].innerText
-    var p = "<p>Hostname:" + hostname + "</p>"
-    $('.modal-body').append(p)
-    var url = $(this).parent('td').siblings('input')[0].value
-    $('#confirm-delete').attr({"data-delete-url": url})
-}
-
-function confirmDelete() {
-    var url = $(this).attr("data-delete-url")
-    $.ajax({
-        url: url,
-        method: 'DELETE'
-    }).done(function (data, status, xhr) {
-        if (xhr.status == 204) {
-            alert('Delete Success.')
-            window.location.reload()
-        }
-    }).fail(function (err) {
-        alert('Oops, something wrong, please contact your Administrator.')
-    })
 }
 
 // ajax: accept minion key
@@ -73,7 +47,8 @@ function acceptMinion() {
     })
 }
 
-//execute_command.html
+// execute_command.html
+// button execute
 function execute(event) {
     let tab = event.data.tab
     let tgt_type = event.data.tgt_type
@@ -125,6 +100,23 @@ function execute(event) {
     })
 }
 
+$('#glob-execute').click({
+    url: "{% url 'api-deploy:execute-cmd' %}",
+    tab: 'glob',
+    tgt_type: 'glob'
+}, execute)
+$('#glob-add').click({tab: 'glob'}, add)
+$('#glob-minus').click({tab: 'glob'}, minus)
+
+$('#server-execute').click({
+    url: "{% url 'api-deploy:execute-cmd' %}",
+    tab: 'server',
+    tgt_type: 'list'
+}, execute)
+$('#server-add').click({tab: 'server'}, add)
+$('#server-minus').click({tab: 'server'}, minus)
+
+// button add
 function add(event) {
     $(`#${event.data.tab}-cmd`).data({'count': $(`#${event.data.tab}-cmd`).data().count + 1})
     let count = $(`#${event.data.tab}-cmd`).data('count')
@@ -132,6 +124,7 @@ function add(event) {
     $(`#${event.data.tab}-arg`).append(`<input type='text' class='form-group form-control' id='${event.data.tab}-arg-${count}' />`)
 }
 
+// button minus
 function minus(event) {
     let count = $(`#${event.data.tab}-cmd`).data('count')
     if (count > 1) {
@@ -140,3 +133,44 @@ function minus(event) {
         $(`#${event.data.tab}-cmd`).data({'count': count - 1})
     }
 }
+
+// salt_ssh.html
+$(".file-dropzone").dropzone({
+    maxFiles: 10,
+    maxFilesize: 5120,
+    acceptedFiles: ".roster"
+});
+
+// button delete roster
+function delRoster(ele, url, fileName) {
+    let modalBody = `<p>File Name: ${fileName}</p>`
+    del(ele, url, modalBody)
+}
+
+// delete modal button confirm
+$('#confirm-delete').click(confirmDel)
+
+// button install minion
+function installMinion(event) {
+    let _tab = event.data.tab
+    let _pk = $('#roster').val()
+    $.ajax({
+        url: `/api/deploy/install-minion/${_pk}/`,
+        method: 'GET'
+    }).done(function (data, status, xhr) {
+        if (xhr.status == 200) {
+            for (let key in data) {
+                $(`#${_tab}-feedback`).append(`<h4>${key}:</h4>`)
+                for (let _key in data[key]) {
+                    if (typeof(data[key][_key]) == "object") {
+                        $(`#${_tab}-feedback`).append(`<p>${_key}: ${JSON.stringify(data[key][_key])}</p>`)
+                    } else {
+                        $(`#${_tab}-feedback`).append(`<p>${_key}: ${data[key][_key]}</p>`)
+                    }
+                }
+            }
+        }
+    })
+}
+
+$('#minion-install').click({tab: 'install'}, installMinion)